requirements.txt
fastapi
uvicorn
pillow
rembg
transformers
diffusers
torch
accelerate
onnxruntime
onnxruntime-gpu
numpy==1.26.4
python-multipart